// src/features/prices/enums/cryptocurrencies.enum.ts
export enum Cryptocurrencies {
  BTCUSDT = 1,
  ETHUSDT = 2,
  XRPUSDT = 3,
  BNBUSDT = 4,
  SOLUSDT = 5,
  DOGEUSDT = 6,
  TRXUSDT = 7,
  ADAUSDT = 8,
  SUIUSDT = 9,
  XLMUSDT = 10,
  LINKUSDT = 11,
  HBARUSDT = 12,
  BCHUSDT = 13,
  AVAXUSDT = 14,
  LTCUSDT = 15,
  TONUSDT = 16,
  SHIBUSDT = 17,
  UNIUSDT = 18,
  DOTUSDT = 19,
  ENAUSDT = 20,
  PEPEUSDT = 21,
  AAVEUSDT = 22,
  TAOUSDT = 23,
  ETCUSDT = 24,
  NEARUSDT = 25,
  APTUSDT = 26,
  ONDOUSDT = 27,
  ICPUSDT = 28,
  ARBUSDT = 29,
  POLUSDT = 30,
  ALGOUSDT = 31,
  PENGUUSDT = 32,
  VETUSDT = 33,
  ATOMUSDT = 34,
  BONKUSDT = 35,
  RENDERUSDT = 36,
  SEIUSDT = 37,
  TRUMPUSDT = 38,
  FILUSDT = 39,
  FETUSDT = 40,
  JUPUSDT = 41,
  TIAUSDT = 42,
  FORMUSDT = 43,
  IOTAUSDT = 44,
  OPUSDT = 45,
  CRVUSDT = 46,
  QNTUSDT = 47,
  STXUSDT = 48,
  LDOUSDT = 49,
  FLOKIUSDT = 50,
}